+++ domains/bank/src/main/kotlin/de/jfachwert/bank/BLZ.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2020 by Oliver Boehm
+ * Copyright (c) 2017-2026 by Oliver Boehm
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -79,6 +79,10 @@ open class BLZ
             }
             return buf.toString().trim { it <= ' ' }
         }
+
+    override fun isValid(): Boolean {
+        return VALIDATOR.isValid(code)
+    }
 
     /**
      * Dieser Validator ist fuer die Ueberpruefung von BLZs vorgesehen.