+++ domains/post/src/test/java/de/jfachwert/post/NameTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 by Oliver Boehm
+ * Copyright (c) 2019-2026 by Oliver Boehm
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -192,6 +192,11 @@ public final class NameTest extends AbstractFachwertTest<String, Text> {
     public void testToShortString() {
         String mozart = "Wolfgang A. Mozart";
         assertEquals(mozart, Name.of(mozart).toShortString());
+    }
+
+    @Test
+    public void testInvalid() {
+        assertFalse(Name.NULL.isValid());
     }
 
 }