+++ CHANGELOG.md
@@ -9,10 +9,11 @@ Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Pa
 ## [Planned]
 - Gradle-Files überarbeiten (Deprecated-Aufrufe raus)
 - auf Gradle 9 und Kotlin 2.3 aktualisieren
-- eigenes validation-Modul für Support von Jakarta-Validation
 
 ## [Unreleased]
 ### Added
+- eigenes validation-Modul für Support von Jakarta-Validation
+- SBOM-Unterstützung
 - validateStrict(..)-Methode im IK.Validator, die fuer alle Klassifikationen die Pruefziffer ueberprueft
 
 ## [6.7.1] - 2026-07-02